projectile/requirements.txt
aniso8601==7.0.0
Django==2.0.5
django-cors-headers==2.5.3
djangorestframework==3.9.3
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
Pillow==6.0.0
promise==2.3
six==1.14.0
